made utility & ecosystems lessons into previewable state
