+++ src/app/econ/utility-optimization/lesson/UtilityOptimizationLesson.tsx
