+++ src/app/environmental-science/ecosystems/lesson/EcosystemExplorer.tsx
