+++ src/app/environmental-science/ecosystems/lesson/PopulationSimulator.tsx
